// src/app/api/model/ruleDataResponseDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { RuleInputResponseDTO } from './ruleInputResponseDTO';
import { RuleOutputResponseDTO } from './ruleOutputResponseDTO';

export interface RuleDataResponseDTO { 
    inputData?: Array<RuleInputResponseDTO>;
    outputData?: Array<RuleOutputResponseDTO>;
}